# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Genesis Trade Suite API
  version: 1.0.0
  description: >
    Production REST API for Genesis Trade Suite, providing Shopify-native
    trade accounts, real-time credit controls, double-entry AR ledger receipts,
    multi-document payment allocations, and eight-bucket aging analysis.
servers:
  - url: https://api.genesistradesuite.com/v1
    description: Production API
  - url: http://127.0.0.1:3001/v1
    description: Local development API
security:
  - ShopifySessionToken: []
paths:
...
  /allocations:
    post:
      summary: Create document allocation
      description: Allocates open credit documents against open debit invoices using explicit or oldest-first mode.
      operationId: createAllocation
      tags:
        - Allocations
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocationRequest'
      responses:
        '201':
          description: Allocation executed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /allocations/{id}/reverse:
    post:
      summary: Reverse document allocation
      description: Appends an immutable allocation reversal, restoring document balances without mutating history.
      operationId: reverseAllocation
      tags:
        - Allocations
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the allocation to reverse.
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReverseAllocationRequest'
      responses:
        '200':
          description: Allocation reversed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReverseAllocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
components:
  securitySchemes:
    ShopifySessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Signed Shopify App Bridge session token HMAC-SHA256 verified.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique idempotency key preventing duplicate command execution.
      schema:
        type: string
  schemas:
...
    AllocationRequest:
      type: object
      required:
        - debtorAccountId
        - mode
        - effectiveDate
      properties:
        debtorAccountId:
          type: string
          format: uuid
        mode:
          type: string
          enum: [explicit, oldest_first]
        effectiveDate:
          type: string
          format: date
        debitDocumentId:
          type: string
          format: uuid
          description: Required when mode is 'explicit'.
        creditDocumentId:
          type: string
          format: uuid
          description: Required for both modes.
        amount:
          type: string
          description: Required when mode is 'explicit'.
    AllocationResponse:
      type: object
      required:
        - mode
        - debtorAccountId
        - unappliedCredit
        - netBalance
      properties:
        mode:
          type: string
          enum: [explicit, oldest_first]
        allocationId:
          type: string
          format: uuid
        debtorAccountId:
          type: string
          format: uuid
        debitDocumentId:
          type: string
          format: uuid
        creditDocumentId:
          type: string
          format: uuid
        amount:
          type: string
        totalAllocated:
          type: string
        unappliedCredit:
          type: string
        netBalance:
          type: string
        isDuplicate:
          type: boolean
        allocations:
          type: array
          items:
            type: object
            properties:
              allocationId: { type: string, format: uuid }
              debitDocumentId: { type: string, format: uuid }
              amount: { type: string }
    ReverseAllocationRequest:
      type: object
      required:
        - reason
        - effectiveDate
      properties:
        reason:
          type: string
        effectiveDate:
          type: string
          format: date
    ReverseAllocationResponse:
      type: object
      required:
        - reversalId
        - allocationId
        - restoredAmount
        - effectiveDate
        - reason
        - unappliedCredit
        - netBalance
      properties:
        reversalId:
          type: string
          format: uuid
        allocationId:
          type: string
          format: uuid
        restoredAmount:
          type: string
        effectiveDate:
          type: string
          format: date
        reason:
          type: string
        unappliedCredit:
          type: string
        netBalance:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
        - code
        - message
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Missing required fields or malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or missing valid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated actor lacks required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found for this tenant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Concurrency version conflict, duplicate identifier, or idempotency conflict with differing payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Business policy rejection, credit declined, or currency mismatch.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'